Fix Xcode 6.3 complaint

+++ KFTunableSpec.m
@@ -732,7 +732,7 @@ static void DrawCloseArtwork(CGContextRef context, CGRect bounds)
     CGFloat components[4];
     
     transform = CGContextGetUserSpaceToDeviceSpaceTransform(context);
-    resolution = sqrtf(fabsf(transform.a * transform.d - transform.b * transform.c)) * 0.5f * (bounds.size.width / imageBounds.size.width + bounds.size.height / imageBounds.size.height);
+    resolution = sqrtf(fabs(transform.a * transform.d - transform.b * transform.c)) * 0.5f * (bounds.size.width / imageBounds.size.width + bounds.size.height / imageBounds.size.height);
     
     CGContextSaveGState(context);
     CGContextClipToRect(context, bounds);
